Restructure public protocols
